Change: split INetworkService into IClientNetworkService and IServerNetworkService
They do distinct things, it makes more sense to use them like this. There was only a couple Services that depended on a generic INetworkService to grab the playerId. I moved the playerId to be managed by IPlayerService instead.

I also disabled the hacky ass block temporary placing because it crashes LocalHost LOL